split ingress tests into multiple files
